Add next-pointer to coap_periodic_resource_s
Items of type `coap_periodic_resource_s` might end up as parts of a linked
list. However, the struct was missing a next pointer field, which meant
that other fields of the struct was overwritten when the item was put in
a list using `list_add`. See `coap_resource_s` for reference.