lc-addrlabels: remove statement expression (fixes #986)
According to GCC documentation, jumping into a statement expression with
a computed goto (using labels as values) has undefined behavior and
should be avoided.  However, this is exactly what is happening in the
current LC_SET implementation.

Fortunately, the statement expression which contains the label in LC_SET
can safely be removed since the operand is only evaluated once, and
multi-statement operations of the macro are already handled by a
do-while(0) construct.