tsch: introduce priority enqueueing (PACKETBUF_ATTR_PRIORITY)
If the first packet in the TX queue is a normal packet and has not
been transmitted ever, a priority packet can be enqueued as a
replacement of the normal packet.

At this moment, a priority packet queue will be enqueued as the last
of the TX queue when the first packet doesn't meet the conditions
mentioned above.