sixtop: increment next_seqno immediately after enqueueing a new request
What RFC 8480 describes for seqno management could cause the initiator
to have two consecutive requests having the same seqno, which would
bring a confusion to a SF if a response to the first request arrives
after the initiator sends the second one. That delayed response will
be treated as a legitimate response to the second request, but it is
not true.

To prevent that, next_seqno is incremented just after having a request
ready to send.