lib: memb: simplify handling of allocation status
During formal verification of memb* functions it was discovered that
reference counting in the functions could be significantly simplified.

Here is the citation from the work:
"This verification case study also allowed to detect a potentially
harmful situation. The memb_free function used to decrement the count
associated to the given block, instead of setting it to 0. An awkward
consequence of this is that calling memb_free on a block with an unusual
count (e.g., greater than 2) would not actually free it. While this
should not happen under normal circumstances, we have decided to replace
that decrement operation by a set to 0. This choice makes memb_free both
simpler and more robust, easing the verification process..."

"Formal Verification of a Memory Allocation Module of Contiki with
Frama-C: a Case Study" Frédéric Mangano, Simon Duquennoy, Nikolai Kosmatov

This commit adopts the verification results from the work.

Signed-off-by: Denis Efremov <efremov@linux.com>
Signed-off-by: Nikolay Kosmatov <nikolaikosmatov@gmail.com>
Signed-off-by: Allan Blanchard <allan.blanchard@inria.fr>
Signed-off-by: Frederic Loulergue <frederic.loulergue@nau.edu>