Fixed a Memory Management bug in the MPL Driver
There was a bug in the MPL driver that could cause the references to the
minimum sequenced numbered message to get out of sync, resulting in old
messages being accepted inserted into the list in the wrong place. This commit
both fixes that issue, and improves the MPL Control message generation
to prevent it entering an infinite loop should the list be out of order.