[bugfix] Disable SysTick interrupt when dropping to some low-power mode
It was discovered that the SysTick interrupt must be disabled before
dropping to some low-power mode. If not, then in some rare occasions a
race condition of the device entering/leaving the low-power mode and the
SysTick interrupt triggering may clobber the CPU context and in an
invalid state, and will in most cases crash the CPU.

A "proxy" standby policy is implemented which wraps the actual Power
driver standby policy with disabling and enabling the SysTick interrupt,
before and after. This ensure the SysTick is always disabled when
entering some low-power mode, regardless of which context that triggers
the standby policy.