Store message payload when using DTLS in coap_receive
The confirmable type needs to be acknowledged. When DTLS is used the ack can no
longer be contained only in the header part of the coap message and the call to
coap_sendto() overwrites the payload of a message stored in uip_buf (calls
uip_udp_packet_sendto() from dtls_write with a data length > 0), hence the need
to store the payload before sending the ack is needed.