Delete the obsolete very-sleepy-demo
This example was first developed for the original Contiki OS and was designed
under the assumption that the underlying radio duty cycling is controlled by
ContikiMAC.

This example has not been updated since the release of Contiki-NG and was only
provided here for reference. In hindsight, its presence in the source tree is confusing
users and overall does more harm than it does good.

This commit removes it.