Native: Wake up 1 time per second instead of 1000
According to the comment about SELECT_TIMEOUT in platform.c for the
native platform it is used to configure the maximum time the scheduler
is allowed to sleep. The unit is specified to be milliseconds.

However, in reality the define was interpreted as number of microseconds
instead of milliseconds. The default value of 1000 made applications
wake up at least 1000 times per second instead of 1 when running
natively.

This fixes that bug and hopefully saves at least some CPU cycles.