Change what text files we consider for the doxygen build
Currently we consider all *.txt files as candidates of containing doxygen documentation. This list of files is getting unnecessarily long, especially as a result of various text files shipped as part of our various submodules. This commit changes our doxygen build to only consider files with a name of a `doxygen*.txt` pattern.