tests: ensure cd is successful
Running 19-out-of-tree-build in the Docker
image on my local computer hangs without
output. The reason is that the return value
of cd is ignored in the Makefile, replacing
";" with "&&" gives us a proper failure.
The clean rule has a similar problem.

Getting the message that make could not cd
into /home/user/out-of-tree-tests can still
be confusing, so add an explanation for
the failure by adding a dependency on the directory
and make the rule for that directory echo
the expected action and then fail.

The reason this test works in the Github CI
is that build.yml ensures /home/user/out-of-tree-tests
is populated before starting docker.