tests: display cooja output
Let the Cooja output echo to stdout in
the tests. This simplifies simexec.sh
and also gives a slight speedup since
the removed polling loop always slept
for 1 second per iteration.

This requires some updates to how test
04-test-result-visualization.sh in 06-script-base
handles its output. Also add a missing "TEST OK"
printout at the end of the script to get the test
included in the summary.