tests: add missing clean
This fixes a link failure in 02-ringbufindex
because of mismatched object types when running
tests in the 32 bit docker and a 64-bit host
using the same source tree.

Add the make clean for the sky/z1 tests as well
to reduce the risk that future cut&paste based
on existing tests reintroduces this issue.