cooja: use C99 uintptr_t
Stop assuming a 32 bit platform when __WORDSIZE
is missing and use C99 INTPTR_MAX to detect
the size of pointers. Fail with an error message
when detection fails.

Unsigned long will not fit a register on
all platforms so switch the types in
the frame struct to use uintptr_t. There is
no general guarantee uintptr_t is sufficient,
but the rest of the code uses inline assembly
for X86 and this will work on all modern
X86 platforms.