Makefile: adjust link rule for Cooja compat
Make the link rule output to $(LIBNAME) and
set $(LIBNAME) to $@ which was the previous
output name.

The "interface" of passing target names through
an environment variable like Cooja does is not
ideal, but revising that interface will be less
cumbersome when the responsibility of the build
system belongs to Contiki-NG and Cooja is a regular
caller.

This commit does not change any visible behavior
for any platform. The purpose is to prepare for
making Cooja use the same build rules as other
platforms.