Makefile: only include *.d for deps
This adjusts the build rules to mostly match
the blog post that was linked in the initial
commit 7a228fea41f7. There are two adjustments
compared to the blog post:

1) Our SOURCEFILES can contain sources in any
   language, so we need to filter out the ones
   that can get automatic dependencies generated.
2) No dependency on $(DEPDIR)/%.d for the object
   build rules because that changes the build
   for Cortex-M due to a .SECONDEXPANSION.

Before this patch, make would try to include
$(OBJDIR)/%.c as Makefiles for dependencies.
There should never have been any c files in
$(OBJDIR), but this behavior might explain
why the CUSTOM_RULE_C_TO_O does not have
dependency generation.

The lack of dependency on $(DEPDIR)/%.d is an issue,
but things are no worse than before this patch.