CI: pass CI variables to docker
Makefile.include contains special logic
to be verbose when the CI environment
variable is true, but we never pass the special
environment variables on to docker.

Rather than trying to decide what CI variables
that might be important for us in the future,
pass all the variables shown in the picture
at the bottom of:

https://docs.github.com/en/actions/creating-actions/creating-a-docker-container-action#testing-out-your-action-in-a-workflow

This change gives the expected CI output when
building tests in Cooja.