Makefile: remove $(CONTIKI_NG_TARGET_LIB)
The -Wl,-zdefs flag when linking only applies
to the object files, not the libraries. Using
the object files directly enables the check
for core OS object files as well.

An additional benefit from the change is that
the build performance improves slightly,
the archive file (which is basically a copy
of the object files) does not have to be
generated/regenerated.

This change also brings the build system
closer to being able to provide an option
for link-time-optimization in the future.