tests/examples: fix TARGET for clean and add -j
Add -j$(CPUS) to the Cooja make commands to speed up
builds.

Discovered that some tests clean for the wrong
TARGET while changing these files, so fix that
at the same time.

The tests all make clean to ensure there is nothing
laying around in the build tree from earlier tests,
and the example result-visualization is used by
the tests so add a make clean in that file.