dbg-io: add format attribute to format_str_v
The format_str function declared just before
this function has the attribute so add it
to format_str_v for consistency. The GCC
documentation explicitly mentions using 0
as the third argument to the format attribute
for va_list functions such as vprintf.

The attribute does not change code generation,
it just informs the compiler that -Wformat checking
should be applied to the arguments of the function.