arm: remove -fno-builtin
The flag entered the repository through
the nrf52832 port. Why that flag was added
to the nRF5 SDK in the first place is unclear:

https://devzone.nordicsemi.com/f/nordic-q-a/53021/reason-for--fno-builtin-introduced-in-sdk-7

The flag is a sledgehammer, -fno-builtin-printf
disables -Wformat warnings, and optimizations
for general purpose functions such as strcmp/memcmp,
general purpose floating point functions,
and typical bit operation functions such as
clz/ctz/popcount.

Removing this flag triggers a warning in cc26x0-web-demo,
so disable the warning for that example so the compiler
can warn for the rest of the code base.