cooja: remove -fno-builtin-printf
The original motivation to add the flag was
probably to prevent compiler optimizations
that replace calls to printf with other
functions, for example puts.

Since the Cooja platform now uses dbg-io,
the concern about compiler optimizations
is no longer an issue so remove the flag
so we get -Wformat warnings from the compiler.