tests: show output in 17-tun-rpl
The intermittent CI failures of 17-tun-rpl
are difficult to debug without relative ordering
of events. Stop writing output to a file that
is shown on failure so we can use
"Show timestamps" in CI to understand the order
of events.

I cannot find any CI logs where only
some pings go through, so remove that
logic as well to eliminate other sources
of errors.