Makefile: autodetect $(CONTIKI)
The build system only verifies that
CONTIKI is set, there is no sanity
check on what it is pointing to. Set
CONTIKI to the directory where
Makefile.include resides.

This eliminates the possibility of standing
in a Contiki-NG v4.7 source tree and pointing
CONTIKI to a v4.6 source tree to use that
build system instead.