ipv6: remove duplicate uip_chksum declaration
The function is already declared in net/ipv6/uip.h
which is included at the top of the header.
This is not the only duplicate declaration in this
header file, but this is the only declaration that
causes a Doxygen warning so remove this and leave
the others for after the release.