CI: add ccache action
This adds the ccache action to CI runs.
The directory is mounted under /home/user/.ccache,
which is how the ccache action configures things.

Some speedups in CI on a warm cache (before/after):

01-compile-base: 1m36s/10s
02-compile-arm-ports: 11m3s/2m30s
07-simulation-base: 4m35s/3m41s
17-tun-rpl-br: 12m48s/12m51s

The first two tests only compiles, and they
both get >99.9% hitrate for the cache so the speedup
is significant.

The third test compiles and performs some work,
so the 100% hitrate for the cache gives a smaller
speedup.

The fourth test gets a 100% hitrate for the cache,
but most of the time is spent on other things.
The increased time for the second run is not
caused by ccache, most likely the second
runner was a slower machine.