Makefile: compile without builtin rules
GNU Make 4.0 made it possible to specify
the --no-builtin-rules inside the Makefile
through MAKEFLAGS.

Remove the existing attempts to disable
builtin rules and use this mechanism instead.
Also add a version check so we get a nice error
message if we are using an old version of make.

The tests are built exctly the same as before,
but running "make -d" reveals that this change
eliminates a lot of attempts to match the builtin
rules for SCCS.