nrf52832: remove $(NRFLIB)
Targets that have their own way of constructing
the list of object files to be built cause issues
with completing the dependency management in
the build system.

Since all files are compiled with the same flags,
there is no strong reason to "partition" the build
and hamper build parallelism. Add the source files
from the SDK to CONTIKI_SOURCEFILES instead of
constructing something special for this architecture.

All object files are still built with the same flags
as before this patch, the only thing that has changed
is that the object files are passed directly to the linker
instead of first making a library that is passed to the linker.