Makefile: remove .PRECIOUS
The .PRECIOUS directive keeps files
even when the command that produced
them was interrupted and similar error
exit conditions. Remove the precious
since we now have .SECONDARY instead,
which is a much smaller sledge hammer.

The output of

env RELSTR=test make NUMCORES=1 Q=''

for 02-compile-arm-ports is identical
before/after this patch.