Dockerfile: change version of traceroute6
More recent versions of iptuils-tracepath
do not contain traceroute6. The traceroute
package is the obvious replacement, but
that gives "* * *" as one or several results
in 17-tun-rpl-br.

The traceroute6 in traceroute most likely
uses a different way of probing hosts and
that doesn't work with Cooja and/or some
component in the regression tests.

Rather than diagnosing the exact issue,
install mtr-tiny and use that since it
works out of the box for the regression
tests.