tests: return error from make for script tests
Mirror the logic from Makefile.simulation-test
so make exits with an error upon failure.

There is no way to make a small change to
the individual tests to adapt to this structural
change, so make this change easy to review by
letting some tests exit with an error code
on first failure.

This commit will enable a significant reduction
of the lines of code in shell scripts under tests.