cortex-m: stop using $(TARGET_LIBS)
The build system does not make a difference
between the variables that are prerequisites
for the link rule. Remove the nrf52832-specific
TARGET_LIBS and use CONTIKI_OBJECTFILES instead.
This is one of the last steps before the customrules
for cortex-m can be removed.

The output from:

env RELSTR=test make NUMCORES=1 Q=''

on 02-compile-arm-ports is identical modulo ordering
before/after this change.