CI: reduce fetch-depth
Reduce the fetch-depth from the entire
history down to 2 now that the Docker
image hash is computed based on file
content rather than the git hash.

A fetch-depth of 2 should be enough to
detect which files that were changed
in a PR, for example for running additional
checks such as code-formatting or similar
on them.