Makefile: unify debug info generation
Add a DEBUG configuration of the build system
that defaults to off. Also provide a default
setting that works with GCC-compatible compilers.
Targets can override both settings depending
on their needs.

MSP430 currently generates stabs with gdb
extensions so preserve that behavior for now.

Besides the order of parameters for MSP430 targets,
the output of:

cimake NUMCORES=1 Q=''

for 01-compile-base and 02-compile-arm-ports
is identical before/after this patch.