tests: update .cscs for modern Cooja
This change aligns the seed in the .csc
with what is used in the regression tests,
prefers the shorter [CONFIG_DIR] over [CONTIKI_DIR],
and adds missing sections/removes unused sections.

The change also adds a spurious newline at the end
of the files, but removing that by hand is not
worth the effort since Cooja will add it back
the next time the files are updated.