Makefile: start Cooja for csc files
This makes it possible to run:

make -f Makefile.include tests/07-simulation-base/01-hello-world-sky.csc

and have Cooja quickstart with a MSP430
simulation.

The "ifndef LIBNAME" was a leftover from
when the LHS of the rule was $(MAKECMDGOALS),
and $(CURDIR) is not required since it is
already defined by GNU make.