tests: update csc files to latest Cooja
Files are now versioned so Cooja can detect
incompatibilities.

This version uses a hierarchical layout for
mote types and their motes, so the mote type
identities are no longer required in the csc
files. This removes the churn when saving
files that have not been updated.

The SimControl plugin has been integrated into
Cooja itself.