cooja: allow building outside Cooja
Removing JNI removed many failure cases from
the build system, so it is now safe to build
binaries targeting Cooja from the shell.
The binaries built from the shell are named
in the same way that they are for other targets.
Cooja will not use these libraries though,
and instead generate a random name that has
not previously been used.