Fix the lifetime value of the default prefix information option
In rpl-lite, the default value for the prefix information lifetime is set to infinity.
Previously, in rpl-classic, this value was not set at all and thus was 0.
A value of 0 does not make sense, since the lifetime of the prefix then would be 0 seconds.
We now do the same thing as rpl-lite.