nrf5340: preprocess linker script
This runs the linker script through the preprocessor.
There are only constants in the linker script right
now so the only thing that changes is the location
of the linker script when invoking the final link.

This is preparation for adding TrustZone support
where the secure and non-secure areas must be
at separate addresses.