tests: stop producing unused logs
The build system regulates how much information
that is printed when compiling, so stop piping
output to files when possible. This provides
timestamps in the CI logs.

Also add -j4 on calls to make to get some
parallelism, and stop tests early
with either executing the shell with -e
or adding "|| exit 1" to the make commands.