Makefile.compile-test: add Clang warning check
Add a second round of (optional) tests using
Clang. This will allow the CI to ensure that
code builds without warnings with Clang.

This commit is just the test infrastructure
support, there are still quite a few PRs with
warning fixes to be merged before the check
can be enabled in the CI.

This commit also makes it possible to run
the tests with:

make Q=''

and see the compiler invocations the same way
that one would when building an example.