cooja: allow disabling devices
This makes it possible to build ContikiMotes
without code that handles pir/vib sensors,
rs232, and so on.

This mechanism is not intended for manual use,
Cooja will call the build system with the right
parameters based on configured mote interfaces
for the mote type.

Size difference between regular hello-world and
all switches except COOJA_RADIO set to 0:
  text	   data	    bss	    dec	    hex
 111352	   4568	  48120	 164040	  280c8
  96013	   3624	  40880	 140517	  224e5

So this saves ~8k * 2 copying per tick, for every
Contiki mote.